experiments/requirements.txt
backports.ssl-match-hostname==3.4.0.2
ecdsa==0.11
hapy-heritrix==0.1.1
matplotlib==1.3.1
nose==1.3.0
numpy==1.8.0
pandas==0.13.1
paramiko==1.12.2
psutil==2.0.0
pycrypto==2.6.1
pyparsing==2.0.1
python-dateutil==2.2
pytz==2013.9
requests==2.2.1
sh==1.09
six==1.5.2
tornado==3.2
wsgiref==0.1.2
